validation: separate manifest and command contracts
